Use the Docker container name without linked parent container names
All linked container names contain the parent and child container name
(the child container name as it's seen by the linked parent, which is
not necessarily the same as the child container's name as seen by
Docker). The linked containers have two forward slashes, so the child
container name as seen by Docker is the only one with a single forward
slash.

Closes #62